# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Core backend service for BookEat.\n\nEvery response is wrapped in a JSON envelope — {\"data\": ...} on\nsuccess, {\"error\": \"...\"} on failure — except /.well-known/jwks.json,\nwhich returns a raw JWKS document.",
        "title": "BookEat backend-core API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the profile of the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/users.userResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft-deletes and anonymizes the authenticated user's account.\nBookings/payments keep their reference to the (now anonymized)\nuser id; they are never deleted or altered. Idempotent.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete current user's account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially updates the authenticated user's profile. Only the\nprovided fields are changed; omitted fields are left untouched.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update current user",
                "parameters": [
                    {
                        "description": "Fields to update (all optional)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.updateMeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/users.userResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "users.updateMeRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/a/jane.png"
                },
                "birth_date": {
                    "description": "BirthDate is a plain date, \"YYYY-MM-DD\".",
                    "type": "string",
                    "example": "1998-05-04"
                },
                "city": {
                    "type": "string",
                    "example": "almaty"
                },
                "country_code": {
                    "description": "CountryCode is an ISO 3166-1 alpha-2 code, e.g. \"KZ\".",
                    "type": "string",
                    "example": "KZ"
                },
                "cuisine_category_ids": {
                    "description": "CuisineCategoryIDs replaces the foodie profile's picked cuisines when\npresent (including an empty array, which clears all picks). Omit the\nfield entirely to leave existing picks unchanged.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "full_name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "preferred_language": {
                    "type": "string",
                    "example": "ru"
                }
            }
        },
        "users.userResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/a/jane.png"
                },
                "birth_date": {
                    "type": "string",
                    "example": "1998-05-04"
                },
                "city": {
                    "type": "string",
                    "example": "almaty"
                },
                "country_code": {
                    "type": "string",
                    "example": "KZ"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-01-15T08:30:00Z"
                },
                "cuisine_category_ids": {
                    "description": "CuisineCategoryIDs is the foodie profile: ids from the restaurants' own\ncategory dictionary (restaurant_categories) the user picked.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "email_verified_at": {
                    "type": "string",
                    "example": "2026-07-10T09:00:00Z"
                },
                "full_name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "phone": {
                    "type": "string",
                    "example": "+77011234567"
                },
                "phone_verified_at": {
                    "type": "string",
                    "example": "2026-07-10T09:00:00Z"
                },
                "preferred_language": {
                    "type": "string",
                    "example": "ru"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token from POST /api/v1/auth/login (or signup / otp/verify /\nrefresh), sent as: Authorization: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}